ISSUE #3967
del: Архивирование оффовского кода
fix: Рефакторинг и устранение ошибок
new: Добавление компонента для погребения
