+++ Content.Shared/_Sunrise/Burial/Components/BuriableComponent.cs
@@ -1,0 +1,10 @@
+namespace Content.Shared.Burial.Components;
+
+/// <summary>
+/// * Компонент позволяющий и/или указывающий на то, что сущность можно закопать в могилу.
+/// </summary>
+[RegisterComponent]
+public sealed partial class BuriableComponent : Component
+{
+
+}
